Added 2nd motor to arm bar, added support for different controller types.
Added a motor to arm bar. Now it has left and right. Also had to update constants for this.

Also made significant updates to robot container to seamlessly accommodate different types of controllers that may be plugged in. Still need to finish updated for radio master controller, and for the "Game Pad" that will be used by the operator.